Add files via upload
All of the necessary files that are needed to 3D print the pieces of the marble toy are .STL files. The .STEP files should be easily readable by any CAD software and will allow you to easily make changes to the parts that have .STEP files. The .SLDPRT and .SLDASM files are Solidworks proprietary 3D files. These files can be converted, however, they are really best used if you have access to Solidworks as they contain the most data and can be changed the easiest.